+++ src/body.rs
@@ -1,16 +1,16 @@
 use std::fmt;
 
 use anyhow::{anyhow, Result};
-use bevy::{prelude::*, time::Stopwatch};
-use derive_more::{Add, From};
+use bevy::prelude::{Color, Component};
+use derive_more::{Deref, DerefMut};
 use genesis_genome::Genome;
 use genesis_util::Probability;
 use rand::RngCore;
 
-use crate::{
-    config,
-    ecosystem::{Energy, Plant},
-};
+use crate::{config, ecosystem};
+
+#[derive(Component, Debug, Deref, DerefMut)]
+pub struct OriginalColor(pub Color);
 
 #[derive(Component, Debug, PartialEq, Eq, Clone)]
 pub struct BugBody {
@@ -18,7 +18,7 @@ pub struct BugBody {
 }
 
 impl BugBody {
-    pub fn new() -> Self {
+    fn new() -> Self {
         let genome = Genome::new(config::CHROMOSOME_COUNT, config::CHROMOSOME_LEN);
 
         Self { genome }
@@ -49,12 +49,12 @@ impl Default for BugBody {
 
 #[derive(Debug)]
 pub struct EnergyReserve {
-    energy: Energy,
+    energy: ecosystem::Energy,
     energy_limit: usize,
 }
 
 impl EnergyReserve {
-    pub fn new(energy: Energy, limit: usize) -> Result<Self> {
+    fn new(energy: ecosystem::Energy, limit: usize) -> Result<Self> {
         if energy.amount() > limit {
             return Err(anyhow!("Limit should be higher than energy passed in."));
         }
@@ -74,24 +74,24 @@ impl EnergyReserve {
     }
 
     #[must_use]
-    pub fn available_space(&self) -> usize {
+    fn available_space(&self) -> usize {
         self.energy_limit - self.energy.amount()
     }
 
     #[must_use]
-    pub fn add_energy(&mut self, mut energy: Energy) -> Energy {
+    fn add_energy(&mut self, mut energy: ecosystem::Energy) -> ecosystem::Energy {
         let energy_taken = energy.take_energy(self.available_space());
         self.energy = self.energy + energy_taken;
         energy
     }
 
     #[must_use]
-    pub fn take_energy(&mut self, amount: usize) -> Energy {
+    fn take_energy(&mut self, amount: usize) -> ecosystem::Energy {
         self.energy.take_energy(amount)
     }
 
     #[must_use]
-    pub fn take_all_energy(&mut self) -> Energy {
+    fn take_all_energy(&mut self) -> ecosystem::Energy {
         let amount = self.amount();
         self.take_energy(amount)
     }
@@ -109,7 +109,7 @@ struct EnergyStore(EnergyReserve);
 struct Health(EnergyReserve);
 
 #[derive(Debug, PartialEq, Eq, Deref, DerefMut)]
-pub struct CoreReserve(Energy);
+pub struct CoreReserve(ecosystem::Energy);
 
 #[derive(Component, Debug)]
 pub struct Vitality {
@@ -119,7 +119,7 @@ pub struct Vitality {
 }
 
 impl Vitality {
-    pub fn new(mut total_energy: Energy) -> Self {
+    pub fn new(mut total_energy: ecosystem::Energy) -> Self {
         let core_energy = config::WorldConfig::global().core_energy;
         let core_reserve = CoreReserve(total_energy.take_energy(core_energy));
 
@@ -156,13 +156,13 @@ impl Vitality {
     }
 
     #[must_use]
-    pub fn add_energy(&mut self, energy: Energy) -> Energy {
+    pub fn add_energy(&mut self, energy: ecosystem::Energy) -> ecosystem::Energy {
         let remaining_energy = self.health.add_energy(energy);
         self.energy_store.add_energy(remaining_energy)
     }
 
     #[must_use]
-    pub fn take_energy(&mut self, amount: usize) -> Energy {
+    pub fn take_energy(&mut self, amount: usize) -> ecosystem::Energy {
         let mut taken_energy = self.energy_store.take_energy(amount);
         let still_needed = amount - taken_energy.amount();
         if still_needed > 0 {
@@ -172,101 +172,19 @@ impl Vitality {
     }
 
     #[must_use]
-    pub fn eat(&mut self, plant: &mut Plant) -> Energy {
+    pub fn eat(&mut self, plant: &mut ecosystem::Plant) -> ecosystem::Energy {
         let requested_energy = self.available_space();
         let extracted_energy = plant.take_energy(requested_energy);
         self.add_energy(extracted_energy)
     }
 
     #[must_use]
-    pub fn move_all_energy(&mut self) -> Energy {
+    pub fn move_all_energy(&mut self) -> ecosystem::Energy {
         let core_energy = self.core_reserve.amount();
         let mut moved_energy = self.core_reserve.take_energy(core_energy);
 
         moved_energy = moved_energy + self.health.take_all_energy();
         moved_energy = moved_energy + self.energy_store.take_all_energy();
         moved_energy
-    }
-}
-
-#[derive(Component, Debug, PartialEq, Eq, Deref, DerefMut, From, Add)]
-pub struct BurntEnergy(Energy);
-
-impl BurntEnergy {
-    pub fn new() -> Self {
-        BurntEnergy(Energy::new_empty())
-    }
-}
-
-impl BurntEnergy {
-    pub fn add_energy(&mut self, energy: Energy) {
-        self.0 = self.0 + energy;
-    }
-
-    pub fn return_energy(&mut self) -> Energy {
-        let amount = self.amount();
-        self.take_energy(amount)
-    }
-}
-
-#[derive(Component, Debug, Deref, DerefMut)]
-pub struct Age(pub Stopwatch);
-
-impl Default for Age {
-    fn default() -> Self {
-        Self(Stopwatch::new())
-    }
-}
-
-impl fmt::Display for Age {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "{}", self.elapsed().as_secs())
-    }
-}
-
-#[derive(Component, Debug, Deref, DerefMut)]
-pub struct Heart(pub Stopwatch);
-
-impl Heart {
-    pub fn new() -> Self {
-        Self(Stopwatch::new())
-    }
-
-    pub fn pulse(&self) -> f32 {
-        self.elapsed_secs().sin()
-    }
-}
-
-#[derive(Component, Debug, Deref, DerefMut)]
-pub struct InternalTimer(pub Stopwatch);
-
-impl InternalTimer {
-    pub fn new() -> Self {
-        Self(Stopwatch::new())
-    }
-}
-
-impl Default for InternalTimer {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl fmt::Display for InternalTimer {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "{}", self.elapsed().as_secs())
-    }
-}
-
-pub fn progress_age_system(time: Res<Time>, mut query: Query<&mut Age>) {
-    for mut age in query.iter_mut() {
-        age.tick(time.delta());
-    }
-}
-
-pub fn progress_timers_system(time: Res<Time>, mut query: Query<(&mut Heart, &mut InternalTimer)>) {
-    for (mut heart, mut internal_timer) in query.iter_mut() {
-        heart.tick(time.delta());
-        internal_timer.tick(time.delta());
     }
 }
