+++ src/ecosystem.rs
@@ -1,8 +1,6 @@
 extern crate derive_more;
-use bevy::prelude::*;
+use bevy::prelude::Component;
 use derive_more::{Add, Constructor, Display, Sub};
-
-use crate::body::BurntEnergy;
 
 #[derive(Clone, Copy, Debug, PartialEq, Eq, PartialOrd, Ord, Add, Display, Sub)]
 pub struct Energy(usize);
@@ -71,15 +69,6 @@ impl Ecosystem {
 
     pub fn return_energy(&mut self, energy: Energy) {
         self.energy = self.energy + energy;
-    }
-}
-
-pub fn burnt_energy_system(
-    mut ecosystem: ResMut<Ecosystem>,
-    mut burnt_query: Query<&mut BurntEnergy>,
-) {
-    for mut burnt_energy in burnt_query.iter_mut() {
-        ecosystem.return_energy(burnt_energy.return_energy())
     }
 }
