+++ src/lib.rs
@@ -1,17 +1,12 @@
 use bevy::prelude::{App, Plugin};
 
 mod attributes;
+mod behaviour;
 mod body;
 mod config;
 mod ecosystem;
-mod interaction;
-mod lifecycle;
 mod mind;
-mod movement;
 mod setup;
-mod sight;
-mod spawn;
-mod systems;
 mod ui;
 
 pub struct GenesisPlugin;
@@ -22,18 +17,16 @@ impl Plugin for GenesisPlugin {
 
         app.insert_resource(config::BACKGROUND)
             .insert_resource(ui::PanelState::default())
-            .add_startup_system(setup::camera_setup)
-            .add_startup_system(setup::physics_setup)
-            .add_system_set(systems::interaction_system_set())
-            .add_system_set(systems::behavior_system_set())
-            .add_system_set(systems::egg_spawning_system_set())
-            .add_system_set(systems::plant_spawning_system_set())
-            .add_system_set(systems::slow_behavior_system_set())
-            .add_system_set(systems::burnt_energy_system_set())
-            .add_system_set(systems::selection_system_set())
+            .add_startup_system_set(setup::setup_system_set())
+            .add_system_set(ui::interaction_system_set())
+            .add_system_set(ui::selection_system_set())
+            .add_system_set(behaviour::behaviour_system_set())
+            .add_system_set(behaviour::egg_spawning_system_set())
+            .add_system_set(behaviour::plant_spawning_system_set())
+            .add_system_set(behaviour::slow_behaviour_system_set())
+            .add_system_set(behaviour::metabolism_system_set())
             .insert_resource(ecosystem::Ecosystem::new(
                 config::WorldConfig::global().world_energy,
-            ))
-            .add_system(ui::energy_ui_update_system);
+            ));
     }
 }
